+++ Assets/Script/PlayerController.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Numerics;
 using UnityEngine;
 using Vector3 = UnityEngine.Vector3;